Add files via upload
Input Format

The first line contains an integer, T, denoting the number of test cases. 
Each test case,T , is comprised of a single line with an integer,n , which can be arbitrarily large or small.

Output Format

For each input variable  and appropriate primitive dataType, you must determine if the given primitives are capable of storing it. If yes, then print:

n can be fitted in:
* dataType